requirements.txt
streamlit-authenticator
streamlit-image-select
websocket-client
python-dotenv